Consolidate crypto_*.go files by capability, not Go type
Five files split by string-aliased type (Crypto, Fingerprint,
Fingerprints, PublicKey, Signature) carried 1-2 methods each and
obscured the actual API: parse PGP, extract identity/signature, verify,
and marshal Fingerprints in/out of GORM. The split made the
security-critical logic hard to find and inflated the file count
without adding a real seam.

Collapse to two files organised by *what the code does*:

- crypto.go — domain types + PGP operations (PublicKey.PGP,
  PublicKey.GetFingerprint, Signature.PGP, Signature.GetSignerFingerprint,
  Signature.Verify, Crypto.Analyze)
- crypto_persistence.go — GORM Scanner/Valuer adapters for Fingerprint
  and Fingerprints

No behaviour changes. go vet, go build, and go test ./... all pass.

Closes #26
